refactor: isolate animation editor state
